Added transition2polynomials.
Perhaps should be renamed, but it finds the polynomials
in a model (given with transition tables, in json format), and
prints the result in json format

Also: code for turning Sook's examples into proper json is given
in read-model.m2.  This file is very rough around the edges, but
hopefully we can get rid of it soon, once json output from the
front end is available.